chore: move SettingsToggleItem
