+++ ui/pages/settings-v2/shared/autodetect-nfts-item.tsx
@@ -8,8 +8,8 @@ import {
   MetaMetricsEventCategory,
   MetaMetricsEventName,
 } from '../../../../shared/constants/metametrics';
-import { SettingsToggleItem } from '../../settings/settings-toggle-item';
 import { ASSET_ITEMS } from '../search-config';
+import { SettingsToggleItem } from './settings-toggle-item';
 
 export const AutodetectNftsToggleItem = () => {
   const t = useI18nContext();
